requirements.txt
fastapi==0.115.0
uvicorn[standard]
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery[redis]==5.4.0
redis==5.0.8
python-dotenv
pytest
httpx